refactor; moved extractors out
