+++ cognite/neat/graph/extractors/__init__.py
@@ -3,9 +3,9 @@ from cognite.neat.graph.extractors._graph_capturing_sheet import (
     extract_graph_from_sheet,
     read_graph_excel_file_to_table_by_name,
 )
+from cognite.neat.graph.extractors._mock_graph_generator import MockGraphGenerator
 
 from ._graph_capturing_sheet import GraphCapturingSheet
-from .mocks._mock_graph_generator import MockGraphGenerator
 
 __all__ = [
     "MockGraphGenerator",
