--- cognite/neat/graph/extractors/mocks/__init__.py
@@ -1,3 +1,0 @@
-from ._mock_graph_generator import generate_triples
-
-__all__ = ["generate_triples"]
